Refactor project, separate game logic from rendering
