+++ src/Game.tsx
@@ -1,0 +1,62 @@
+import { useState } from 'react'
+import { type GameState, initialGameState } from './gamelogic'
+import { makeMove as getNewGameState } from './gamelogic'
+
+const Game = () => {
+  const [gameState, setGameState] = useState(initialGameState)
+
+  const makeMove = (row: number, col: number) => {
+    setGameState(getNewGameState(gameState, row, col))
+  }
+
+  const resetGame = () => {
+    const newGameState = structuredClone(initialGameState)
+    setGameState(newGameState)
+  }
+
+  return (
+    <div>
+      <p>Current player is {gameState.currentPlayer}</p>
+      {gameState.winner ? (<>the winner is {gameState.winner}</>) : ("")}
+      <div className="text-xl text-purple-600">
+        <p>
+          <Cell makeMove={makeMove} gameState={gameState} row={0} col={0}/>
+          <Cell makeMove={makeMove} gameState={gameState} row={0} col={1}/>
+          <Cell makeMove={makeMove} gameState={gameState} row={0} col={2}/>
+        </p>
+        <p>
+          <Cell makeMove={makeMove} gameState={gameState} row={1} col={0}/>
+          <Cell makeMove={makeMove} gameState={gameState} row={1} col={1}/>
+          <Cell makeMove={makeMove} gameState={gameState} row={1} col={2}/>
+        </p>
+        <p>
+          <Cell makeMove={makeMove} gameState={gameState} row={2} col={0}/>
+          <Cell makeMove={makeMove} gameState={gameState} row={2} col={1}/>
+          <Cell makeMove={makeMove} gameState={gameState} row={2} col={2}/>
+        </p>
+      </div>
+      <button onClick={resetGame}>reset game!</button>
+    </div>
+  )
+}
+
+interface CellProps {
+    makeMove: (row: number, col: number) => void,
+    row: number,
+    col: number,
+    gameState: GameState
+}
+
+const Cell = ({ makeMove, row, col, gameState }: CellProps) => {
+  const handleClick = () => {
+    makeMove(row, col)
+  }
+
+  return (
+    <button onClick={handleClick}>
+      {gameState.board[row][col]}
+    </button>
+  )
+}
+
+export default Game
